pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Task Manager on Django"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django",
    "gunicorn",
    "python-dotenv",
    "dj-database-url",
    "psycopg[binary]",
    "django-bootstrap5>=25.1",
    "django-filter==24.1",
    "rollbar==0.16.3",
    "ruff==0.8.1"
]

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = ["migrations", "pycache"]

[tool.ruff.lint]
select = ["E", "F", "I"]  # ошибки, предупреждения, импорт

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 88
terget-version = ['py310']
exlude = '''
/(
    migrations
  | \.venv
  | __pycache__
)/
'''
